arca-sim/pyproject.toml
[project]
name = "arca-sim"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "eclipse-zenoh>=1.3.4",
    "graphviz>=0.21",
    "ipykernel>=6.29.5",
    "meshcat>=0.3.2",
    "msgpack>=1.1.1",
    "mujoco==3.3.2",
    "numpy>=2.2.5",
    "onshape-to-robot",
    "pin>=3.7.0",
    "pymeshlab>=2023.12.post3",
]

[tool.uv.sources]
onshape-to-robot = { git = "https://github.com/Rhoban/onshape-to-robot.git" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/arca_sim"]
